pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["leonidee"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
ipykernel = "^6.25.2"
findspark = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
